chore: move standalone types to types crate
